# openapi.yml
openapi: 3.0.2
servers:
  - url: /v3
info:
  version: 1.0.0
  title: Pet store schema
tags:
  - name: pet
    description: Everything about your Pets
paths:
  /pet:
    post:
      tags:
        - pet
      summary: Add a new pet to the store
      description: Add a new pet to the store
      operationId: addPet
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
      requestBody:
        description: Create a new pet in the store
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                viento_item:
                  $ref: '#/components/schemas/VientoItem'
                report_response:
                  $ref: '#/components/schemas/GetReportResponse'
  # '/pet/{petId}':
  #   get:
  #     tags:
  #       - pet
  #     summary: Find pet by ID
  #     description: Returns a single pet
  #     operationId: getPetById
  #     parameters:
  #       - name: petId
  #         in: path
  #         description: ID of pet to return
  #         required: true
  #         schema:
  #           type: integer
  #           format: int64
  #     responses:
  #       '200':
  #         description: successful operation
  #         content:
  #           application/json:
  #             schema:
  #               $ref: '#/components/schemas/Pet'
  #       '404':
  #         description: Pet not found
  #   post:
  #     tags:
  #       - pet
  #     summary: Updates a pet in the store
  #     description: ''
  #     operationId: updatePet
  #     parameters:
  #       - name: petId
  #         in: path
  #         description: ID of pet that needs to be updated
  #         required: true
  #         schema:
  #           type: integer
  #           format: int64
  #       - name: name
  #         in: query
  #         description: Name of pet that needs to be updated
  #         schema:
  #           type: string
  #       - name: status
  #         in: query
  #         description: Status of pet that needs to be updated
  #         schema:
  #           $ref: '#/components/schemas/PetStatus'
  #     responses:
  #       '200':
  #         description: successful operation
  #   delete:
  #     tags:
  #       - pet
  #     summary: Deletes a pet
  #     description: ''
  #     operationId: deletePet
  #     parameters:
  #       - name: petId
  #         in: path
  #         description: Pet id to delete
  #         required: true
  #         schema:
  #           type: integer
  #           format: int64
  #     responses:
  #       '200':
  #         description: successful operation
  /viento/products:
    post:
      tags:
        - ozon
      summary: get viento products
      description: get viento products
      operationId: getVientoProducts
      responses:
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VientoProduct'
        '400':
          description: bad request

  /ozon/list_transaction:
    post:
      tags:
        - ozon
      summary: get viento list transaction
      description: get viento list transaction
      operationId: getVientoListTransaction
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetSumProficiencyParams'
      responses:
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTransactionParams'
        '400':
          description: bad request

  /ozon/item:
    post:
      tags:
        - ozon
      summary: get ozon items
      description: get ozon items
      operationId: getOzonItems
      responses:
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: bad request

  /ozon/posting:
    post:
      tags:
        - ozon
      summary: get viento posting
      description: get viento posting
      operationId: getVientoPosting
      responses:
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Posting'
        '400':
          description: bad request

  /ozon/service:
    post:
      tags:
        - ozon
      summary: get viento services
      description: get viento services
      operationId: getVientoServices
      responses:
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400':
          description: bad request

  /ozon/operation:
    post:
      tags:
        - ozon
      summary: get viento operations
      description: get viento operations
      operationId: getVientoOperations
      responses:
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '400':
          description: bad request

  /ozon/sum/services:
    post:
      tags:
        - ozon
      summary: get sum services by day
      description: get sum services by day
      operationId: getSumServicesByDay
      requestBody:
        description: Sum services by day parameters
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SumServicesByDayParams'
      responses:
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - name
                    - amount
                  properties:
                    name:
                      type: string
                    amount:
                      type: number
                      format: float
        '400':
          description: bad request

components:
  # Reusable schemas (data models)
  schemas:
    Error:
      description: Represents error object
      type: object
      x-ogen-name: Error
      properties:
        code:
          type: integer
          format: int64
        message:
          type: string
      required:
        - code
        - message
    VientoProduct:
      type: object  
      description: viento product
      x-ogen-name: VientoProduct
      properties:
        ProductID:
          type: integer
          format: int64
          x-oapi-codegen-extra-tags:
            gorm: primaryKey
        OfferID:
          type: string
        IsFBOVisible:
          type: boolean
        Archived:
          type: boolean
        IsFBSVisible:
          type: boolean
        IsDiscounted:
          type: boolean
        Barcode:
          type: string
        Barcodes:
          type: array
          items:
            type: string
        Price:
          type: number
          format: float64
    Item:
      type: object
      description: item in stock
      x-ogen-name: Item
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        sku:
          type: integer
          format: int64
        ean:
          type: integer
          format: int64
        viento_id:
          type: integer
          format: int64
        consumption:
          type: integer
          format: int64
    Posting:
      type: object
      x-ogen-name: Posting
      properties:
        ID:
          type: integer
          format: int64
        DeliverySchema:
          type: string
        OrderDate:
          type: string
        PostingNumber:
          type: string
        WarehouseID:
          type: integer
          format: int64
    Service:
      type: object
      x-ogen-name: Service
      properties:
        ID:
          type: integer
          format: int64
        OperationID:
          type: integer
          format: int64
        Name:
          type: string
        Price:
          type: number
          format: float
    Operation:
      type: object
      x-ogen-name: Operation
      properties:
        ID:
          type: integer
          format: int64
        OperationID:
          type: integer
          format: int64
        OperationType:
          type: string
        OperationDate:
          type: string
        OperationTypeName:
          type: string
        DeliveryCharge:
          type: number
          format: float
        ReturnDeliveryCharge:
          type: number
          format: float
        AccrualsForSale:
          type: number
          format: float
        SaleCommission:
          type: number
          format: float
        Amount:
          type: number
          format: float
        TransactionType:
          type: string
        PostingID:
          type: integer
          format: int64
        ItemSKU:
          type: integer
          format: int64
        Services:
          type: array
          items:
            type: integer
            format: int64
    SumServicesByDayParams:
      type: object
      x-ogen-name: SumServicesByDayParams
      properties:
        date:
          type: string
          format: date-time
    ListTransactionParams:
      type: object
      x-ogen-name: ListTransactionParams
      required:
        - filter
        - filter.operation_type
      properties:
        filter:
          type: object
          properties:
            date:
              type: object
              properties:
                from:
                  type: string
                  format: date-time
                to:
                  type: string
                  format: date-time
            operation_type:
              type: array
              items:
                type: string
            posting_number:
              type: string
            transaction_type:
              type: string
        page:
          type: integer
          format: int64
        page_size:
          type: integer
          format: int64
    GetSumProficiencyParams: 
      type: object
      x-ogen-name: GetSumProficiencyParams
      properties:
        date:
          type: object
          properties:
            from:
              type: string
              format: date-time
            to:
              type: string
              format: date-time

    GetReportResponse:
      type: object
      properties:
        total_count:
          type: integer
          format: int64
          description: общее число проданных единиц
        total:
          type: number
          format: float
          description: общая сумма после коммиссии
        delivery:
          type: number
          format: float
          description: доставка
        refund_and_others:
          type: number
          format: float
          description: возврат
        service:
          type: number
          format: float
          description: сервисный сбор
        final:
          type: number
          format: float
          description: общая сумма
        result:
          type: array
          items:
            $ref: '#/components/schemas/ReportSumInfo'
          description: список проданых товаров по позициям

    ReportSumInfo:
      type: object
      properties:
        item_name:
          type: string
          description: название товара
        count:
          type: integer
          format: int64
          description: количество
        amount:
          type: number
          format: float
          description: сумма
   
    VientoItem:
      type: object
      required:
        - name
        - price
        - ean
        - id
      properties:
        id:
          type: integer
          format: int64
          x-oapi-codegen-extra-tags:
            gorm: primaryKey;autoIncrement
        name:
          type: string
        price:
          type: number
          format: float64
        ean:
          type: integer
          description: barcode from viento price list
          format: int64